IMPALA-12213: Verify if the table exists in the event processor for reload event
When a reload event is being processed in the event processor, it is
ideal to verify if the table exists before checking if the event is
older. In the production use case, it is very likely that create and
reload events for a same table comes in the same batch of events and
this might lead to null pointer exception while processing reload event
since the table does not exist in the cache.

Testing: Added an end-to-end test case to simulate the above scenario
and verify that event processor is in active state at the end of the
test.

Change-Id: I6bf7f596154e3695b7eea723d8cd3b0f7bda83cb
Reviewed-on: http://gerrit.cloudera.org:8080/20117
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>