IMPALA-10048: Go parallel for dump_breakpad_symbols.py
This modifies dump_breakpad_symbols.py to use a ThreadPool
to go parallel when there are multiple binaries or
libraries to process. This is common for Jenkins jobs that
dump symbols for all backend tests. The different binaries
write out to different directories, so the threads don't
interfere with each other.

Testing:
 - Ran locally dumping the symbols for all backend tests
 - Ran a Jenkins job that generates a minidump and triggers
   the minidump symbol processing. It went parallel and
   worked fine.

Change-Id: I93427bb07f1d9718bd6df90acfd247210b54294d
Reviewed-on: http://gerrit.cloudera.org:8080/20802
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>