IMPALA-12248: Add three configuration properties after RANGER-2895
This patch adds 3 Ranger configuration properties that will be required
after we start using a build that includes RANGER-2895 so that Ranger's
HTTP server could be properly started.

On the other hand, recall that a Ranger configuration property was
deprecated in RANGER-2895, i.e.,
ranger.jpa.jdbc.idleconnectiontestperiod. Hence, we should also remove
it when starting using a build that contains RANGER-2895. This task will
be tracked by IMPALA-12250.

Testing:
 - Verified that with this patch Ranger's HTTP server could be started
   and that we will be able to update the users, groups, and policies on
   the Ranger server with the current scripts in the Impala repository
   whether or not we are using a build that has RANGER-2895.

Change-Id: I19a27e3fe3ab96a9f60566dc2c87bd72636b91ae
Reviewed-on: http://gerrit.cloudera.org:8080/20129
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>