IMPALA-12249: Fix the unexpected word wrap of 'Progress' in WebUI queries page
This commit replaces the whitespace in 'Scan Progress' and 'Query
Progress' of queries page with non-breaking spaces (nbsp). This is to
prevent these headers from breaking at the space when displayed in the
browser. With this change, the headers will stay on the same line,
improving the readability of the table headers.

Change-Id: I894ada826282d33c3f2395231db1ddf97bc82367
Reviewed-on: http://gerrit.cloudera.org:8080/20130
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>