IMPALA-12273 Fix possible memory corruption of premature release of objects returned by native methods
In the Jni Native method, there was an erroneous usage of JniLocalFrame.
The JniLocalFrame mistakenly released the reference to the returned
object during its destruction, causing the object to be reclaimed during
high-concurrency garbage collection.

The solution is to remove the JniLocalFrame-related code, since the
native method does not create a local frame in JniEnv.

Testing: During stress testing, there have been no further occurrences
of memory corruption.

Change-Id: I94b1c6b3f96897850512c01cdb942ebdcc59b694
Reviewed-on: http://gerrit.cloudera.org:8080/20174
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>