IMPALA-12274: Fix memory leak from unreleased local reference
In the constructor of the catalog class, after converting the catalog
object created by NewObject to a global reference, the local reference
was forgotten to be released, resulting in a minor memory leak.

The solution is to use JniLocalFrame to automatically manage the local
reference, ensuring proper memory management.

Testing: manually test

Change-Id: Ic93ea3270dcba3ad4903aa053cc283d4f700e948
Reviewed-on: http://gerrit.cloudera.org:8080/20175
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>