IMPALA-12267: DMLs/DDLs can hang as a result of catalogd restart
IMPALA-5476 added handling for changes in catalogd service ID during the
lifetime of a DML/DDL query, but the loop that waits for a new catalogd
service ID can wait indefinitely in case the DML/DDL was handled by the
previous catalogd and restart happened just after replying to the
coordinator:
https://github.com/apache/impala/blob/d0fe4c604f72d41019832513ebf65cfe8f469953/be/src/service/impala-server.cc#L2204

This change adds two new startup flags. The first one is
'--wait_for_new_catalog_service_id_timeout_sec', which sets an upper
limit on the waiting time (in seconds). The second one is
'--wait_for_new_catalog_service_id_max_iterations', which limits how
many valid catalog updates without a new catalog service ID the
coordinator receives before it gives up and stops waiting.

For both startup flags, negative values and zero turn the feature off.

If the coordinator gives up waiting because of either of the new startup
flags, the local catalog cache is not updated and a warning is logged.

Testing:
 - added a custom cluster test that reproduces the above condition using
   a debug action and verifies that the waiting loop times out.

Change-Id: Ib71bec8f67f80b0bdfe0a6cc46a16ef624163d8b
Reviewed-on: http://gerrit.cloudera.org:8080/20192
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>