IMPALA-12286: Make CatalogD HA robust
IMPALA-12155 added support for CatalogD HA. The statestore assigns the
roles for the catalogd in the HA pair. When an active catalogd is
elected, statestore sends RPCs to coordinators and catalogds to notify
the new active catalogd. But the RPCs could fail due to network
conditions. To avoid missing the notification of new active catalogd,
this patch makes statestore to resend RPCs to subscribers if there are
RPC failures.

This patch also rename the metric "catalog-server.ha-active-status" to
"catalog-server.active-status" since this metric is also set for
catalogd when CatalogD HA is not enabled.

As discussed in IMPALA-12267, catalogd need to re-generate its Catalog
Service ID when it becomes active.

Testing:
 - Added unit-test cases for CatalogD HA with simulated RPC failures.
 - Passed core tests.

Change-Id: Ibdfea022031c3cc1cbaf4ad52e947720a5d5630f
Reviewed-on: http://gerrit.cloudera.org:8080/20220
Reviewed-by: Andrew Sherman <asherman@cloudera.com>
Reviewed-by: Abhishek Rawat <arawat@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>