IMPALA-12296: Fix filenames in query profile download page
The downloads of query profiles from the profile page contain ':'
in the filename, this is not supported by some filesystems
such as NTFS.

Most browsers replace it with a white-space during the download.

Hence, characters except letters and white spaces in downloaded
profile's filename should be replaced with '_'.

Tests in 'test_web_pages.py' do not need modifications as
the endpoint for profile downloads has not changed.

Change-Id: Ia170e0ae1adc6f3347b57ca8572d65a05ebf9544
Reviewed-on: http://gerrit.cloudera.org:8080/20215
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>