IMPALA-12292: TestCodegenCache.{test_codegen_cache_with_asm_module_dir,test_codegen_cache_with_perf_map} fail in builds
The above codegen cache tests were introduced by IMPALA-12260. They run
two queries and the first query produces two codegen cache entries. The
tests aim to bring about the following scenario:

1. both codegen cache entries from the first query fit in the cache
AND
2. both entries from the first query are evicted during the second
   query.

The parameters that can be tuned are the following:
1. the size of the codegen cache entries of the first query
2. the size of the codegen cache entries of the second query
3. the size of the codegen cache.

If the parameters are chosen badly or the sizes of the codegen cache
entries change because of other Impala changes (e.g. codegen
optimisations), the conditions may not be satisfied and the tests may
fail like they did now.

This change makes the tests more robust by
 - increasing the cache footprint of the second query (from 487.40 KB to
   663.68 KB)
 - choosing the size of the codegen cache so as to leave as much margin
   on each side as possible. At present
     - the minimal codegen cache size so that both entries from the
       first query fit the cache is around 2.4 MB
     - the maximal cache size so that both entries from the first query
       are evicted during the second query is around 4.1 MB
   Therefore we choose a cache size of 3.25 MB, which lies in the middle.

Experience has shown that this setup is fragile and breaks easily when
new commits are added to Impala. Therefore this change relaxes some of
the assertions in the tests as a temporary measure to prevent build
failures. For this and other reasons IMPALA-12306 was opened to make
these tests more robust.

Change-Id: I15320b8c0d06f4d93927b19731c11bd4e15b3690
Reviewed-on: http://gerrit.cloudera.org:8080/20224
Reviewed-by: Yida Wu <wydbaggio000@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>