IMPALA-12340: Fix flaky test_two_catalogd_with_force_active
The issue could be re-produced by repeatedly running the test case for
about 50 times in Jenkins or on local machine.
The issue was introduced by IMPALA-12304, for which we tried not to
wake up the thread for update_catalogd RPC if there is no change for
elected active catalogd. Since we cannot hold mutex when calling
SendUpdateCatalogdNotification(), it may cause the notification to the
condition variable not been processed. To fix the issue, revert the
optimization added in IMPALA-12304, and add sleep in
test_two_catalogd_with_force_active.

Testing:
 - Repeatedly ran test_two_catalogd_with_force_active for 8000 times on
   local machine without failure, ran the test for 1000 times in Jenkins
   without failure.
   Repeatedly ran all test cases in test_catalogd_ha.py for 21 hours in
   Jenkins without failure.
 - Passed core test.

Change-Id: Ifc213422a5a3360d07a22046e1f46fdf0be1d2fd
Reviewed-on: http://gerrit.cloudera.org:8080/20323
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>